MC2_gesture_detection_2024/requirements.txt
# Execute following commands:
# pip install -r requirements.txt
# conda install pytorch==1.12.1 torchvision==0.13.1 torchaudio==0.12.1 cudatoolkit=11.3 -c pytorch
about-time==4.2.1
absl-py==1.4.0
alive-progress==3.0.1
appdirs==1.4.4
astunparse==1.6.3
black==23.3.0
bottleneck==1.3.5
brotli==1.0.9
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
flatbuffers==23.1.21
flit-core==3.6.0
fonttools==4.25.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grapheme==0.6.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
intel-openmp==2021.4.0
joblib==1.2.0
kaleido==0.2.1
keras==2.11.0
keras-preprocessing==1.1.2
kiwisolver==1.4.4
lazy-loader==0.1
markdown==3.4.1
markupsafe==2.1.2
matplotlib==3.6.3
oauthlib==3.2.2
pandas==1.5.2
pillow==9.3.0
pip==22.3.1
plotly==5.9.0
pydot==1.4.2
pyopenssl==23.0.0
pyparsing==3.0.9
pyserial==3.5
pysocks==1.7.1
pytz==2022.7
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.9.1
seaborn==0.12.2
setuptools==65.6.3
sip==6.6.2
six==1.16.0
tensorboard==2.11.2
tensorflow==2.11.0
termcolor==2.2.0
torchaudio==0.12.1
torchsummary==1.5.1
torchvision==0.13.1
tqdm==4.65.0
# pytorch==1.12.1
# cudatoolkit==11.3.1